awimage: add support for A31 images
This commit makes awimage successfully unpack A31 images (or at least
the one image that was provided to me).

A31 images expose a bug in awimage in imgRepacker mode, namely, if no
full filename is stored, it will only show "Extracting " and not actually
extract the file in question (since it has no name).

This issue will be addressed in another commit.

Signed-off-by: Ithamar R. Adema <ithamar@upgrade-android.com>